requirements.txt
blinker==1.9.0
charset-normalizer==3.4.5
click==8.3.1
contourpy==1.3.3
cycler==0.12.1
dnspython==2.8.0
Flask==3.1.3
fonttools==4.61.1
gunicorn==25.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.3
kiwisolver==1.4.9
MarkupSafe==3.0.3
matplotlib==3.10.8
numpy==2.4.2
opencv-python==4.13.0.92
packaging==26.0
pandas==3.0.1
pillow==12.1.1
pymongo==4.16.0
pyparsing==3.3.2
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
reportlab==4.4.10
scikit-learn==1.8.0
scipy==1.17.1
six==1.17.0
threadpoolctl==3.6.0
Werkzeug==3.1.6
